+++ app/src/main/java/com/techcourse/service/AppUserService.java
@@ -1,13 +1,9 @@
 package com.techcourse.service;
 
-import com.techcourse.config.DataSourceConfig;
 import com.techcourse.dao.UserDao;
 import com.techcourse.dao.UserHistoryDao;
 import com.techcourse.domain.User;
 import com.techcourse.domain.UserHistory;
-import nextstep.jdbc.DataAccessException;
-import org.springframework.jdbc.datasource.DataSourceTransactionManager;
-import org.springframework.transaction.support.DefaultTransactionDefinition;
 
 public class AppUserService implements UserService {
 
@@ -31,19 +27,9 @@ public class AppUserService implements UserService {
 
     @Override
     public void changePassword(final long id, final String newPassword, final String createBy) {
-        var dataSource = DataSourceConfig.getInstance();
-        var transactionManager = new DataSourceTransactionManager(dataSource);
-        var transactionStatus = transactionManager.getTransaction(new DefaultTransactionDefinition());
-        try {
-            var user = findById(id);
-            user.changePassword(newPassword);
-            userDao.update(user);
-            userHistoryDao.log(new UserHistory(user, createBy));
-            transactionManager.commit(transactionStatus);
-        } catch (Exception e) {
-            e.printStackTrace();
-            transactionManager.rollback(transactionStatus);
-            throw new DataAccessException(e);
-        }
+        var user = findById(id);
+        user.changePassword(newPassword);
+        userDao.update(user);
+        userHistoryDao.log(new UserHistory(user, createBy));
     }
 }
